requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
Pillow==10.2.0
python-dotenv==1.0.1
drf-yasg==1.21.7
django-cors-headers==4.3.1
redis==5.0.1
django-redis==5.4.0
django-filter==23.5
django-cleanup==8.1.0
django-storages==1.14.2
celery==5.3.6
gunicorn==21.2.0
black==24.2.0
flake8==7.0.0
mypy==1.8.0
pytest==8.0.2
pytest-django==4.8.0
factory-boy==3.3.0  

# Additional
unidecode==1.3.6 